refactor(infra): move 7 v2 entities + 5 ports to infrastructure/persistence (#896)
Move legacy JPA entities from module-infra/domain/v2/ to
infrastructure/persistence/entity/, port interfaces from
domain/repository/ to infrastructure/persistence/repository/.

Remove business logic from JPA entities:
- Member: deductPoints/hasEnoughPoint/validatePositiveAmount/maskUuid
- GameCharacterV2: like/isActive/needsBasicInfoRefresh/validateOcid
- EquipmentExpectationSummary: updateExpectation/touch
- Replace with atomic queries at port level:
  - MemberRepository.decreasePointByUuid (WHERE point >= amount)
  - MemberRepository.increasePointByUuid
- InternalPointPaymentStrategy uses decreasePointByUuid instead of
  Member.deductPoints (Issue #896 atomic query strategy)

Enum dedup:
- Use module-core CubeType/PotentialGrade
- Drop v2 duplicates; remove v2->core adapter (mapToCoreCubeType)

Caller-side imports updated across:
- module-infra: cube service, repository impls, jpa repository
- module-app: controllers, config, scheduler, donation service
- tests: MemberFixture, MemberRepositoryIntegrationTest,
  JpaNPlusOneRegressionTest, ArchUnit javadoc

Drop dead code:
- CubeTrialsPort (no consumers)
- DomainCharacterizationTest (v2 domain behavior removed)
- MapToCoreCubeType v2->core adapter (now identity)

DB schema unchanged.
