+++ module-app/src/main/java/maple/expectation/application/service/donation/InternalPointPaymentStrategy.java
@@ -3,38 +3,24 @@ package maple.expectation.application.service.donation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import maple.expectation.infrastructure.persistence.repository.MemberRepository;
-import maple.expectation.domain.v2.Member;
 import maple.expectation.error.exception.AdminMemberNotFoundException;
 import maple.expectation.error.exception.SenderMemberNotFoundException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
- * 내부 포인트 시스템 기반 결제 전략 (Hybrid: Rich Domain + Atomic Query)
+ * 내부 포인트 시스템 기반 결제 전략 (Atomic Query)
  *
- * <p>Issue #120: Rich Domain Model 전환
+ * <p>Issue #896: Rich Domain Model 제거, port의 원자적 쿼리로 통일.
  *
- * <h3>하이브리드 전략</h3>
+ * <h3>원자적 쿼리 전략</h3>
  *
  * <ul>
- *   <li><b>Sender (발신자)</b>: Rich Domain (member.deductPoints()) - 각 사용자는 자신의 데이터만 수정하므로 충돌 가능성 낮음
- *       - @Version 낙관적 락으로 보호
+ *   <li><b>Sender (발신자)</b>: 원자적 UPDATE (memberRepository.decreasePointByUuid()) - 잔액 부족 시 0건 반환
+ *       - WHERE point >= :amount 조건으로 Lost Update 방지
  *   <li><b>Receiver (Admin)</b>: 원자적 쿼리 (memberRepository.increasePointByUuid()) - Hot Key 문제: 모든
  *       후원이 단일 Admin으로 집중됨 - 100명 동시 요청 시 원자적 UPDATE로 Lost Update 방지
  * </ul>
- *
- * <h3>포트원 연동 시 교체 방법:</h3>
- *
- * <pre>{@code
- * // application.yml
- * payment:
- *   provider: portone  # internal (기본값) -> portone
- *
- * // PortOnePaymentStrategy.java
- * @Component
- * @ConditionalOnProperty(name = "payment.provider", havingValue = "portone")
- * public class PortOnePaymentStrategy implements PaymentStrategy { ... }
- * }</pre>
  */
 @Slf4j
 @Component
@@ -44,12 +30,10 @@ public class InternalPointPaymentStrategy implements PaymentStrategy {
   private final MemberRepository memberRepository;
 
   /**
-   * 포인트 이체 처리 (Hybrid: Rich Domain + Atomic Query)
-   *
-   * <p>Issue #120: Sender에 Member.deductPoints() 적용 (플랜 핵심 요구사항)
+   * 포인트 이체 처리 (Atomic Query)
    *
    * @param senderUuid 발신자 UUID
-   * @param receiverFingerprint 수신자(Admin) fingerprint (Member.uuid로 사용)
+   * @param receiverFingerprint 수신자(Admin) fingerprint
    * @param amount 이체 금액
    */
   @Override
@@ -60,18 +44,13 @@ public class InternalPointPaymentStrategy implements PaymentStrategy {
         maskUuid(senderUuid),
         amount);
 
-    // 1. 발신자: Rich Domain (Issue #120 핵심)
-    //    - 각 사용자는 자신의 데이터만 수정 → 충돌 가능성 낮음
-    //    - @Version 낙관적 락으로 보호
-    Member sender = memberRepository.findByUuid(senderUuid);
-    if (sender == null) {
+    // 1. 발신자: 원자적 UPDATE (Issue #896: Rich Domain → Port로 이관)
+    //    - WHERE point >= :amount 조건으로 잔액 부족 시 0건 반환
+    if (memberRepository.decreasePointByUuid(senderUuid, amount) == 0) {
       throw new SenderMemberNotFoundException(maskUuid(senderUuid));
     }
-    sender.deductPoints(amount); // InsufficientPointException 발생 가능
 
     // 2. 수신자(Admin): 원자적 쿼리 (Hot Key 보호)
-    //    - 모든 후원이 단일 Admin으로 집중 → 동시성 경합 심함
-    //    - 원자적 UPDATE로 Lost Update 방지
     if (memberRepository.increasePointByUuid(receiverFingerprint, amount) == 0) {
       throw new AdminMemberNotFoundException(receiverFingerprint);
     }
