+++ module-infra/src/main/java/maple/expectation/application/service/cube/component/SlotDistributionBuilder.java
@@ -9,7 +9,7 @@ import maple.expectation.config.TableMassConfig;
 import maple.expectation.core.domain.model.calculator.SparsePmf;
 import maple.expectation.core.domain.stat.StatType;
 import maple.expectation.infrastructure.persistence.repository.CubeProbabilityRepository;
-import maple.expectation.domain.v2.CubeProbability;
+import maple.expectation.infrastructure.persistence.entity.CubeProbability;
 import maple.expectation.core.domain.model.CubeType;
 import maple.expectation.error.exception.ProbabilityInvariantException;
 import maple.expectation.infrastructure.executor.LogicExecutor;
