+++ module-infra/src/main/kotlin/maple/expectation/infrastructure/persistence/repository/MemberRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package maple.expectation.infrastructure.persistence.repository
 
-import maple.expectation.infrastructure.persistence.repository.MemberRepository
-import maple.expectation.domain.v2.Member
+import maple.expectation.infrastructure.persistence.entity.MemberEntity
 import maple.expectation.infrastructure.persistence.jpa.MemberJpaRepository
 import org.springframework.stereotype.Repository
 
@@ -21,14 +20,14 @@ open class MemberRepositoryImpl(
     private val jpaRepository: MemberJpaRepository,
 ) : MemberRepository {
 
-    override fun findByUuid(uuid: String): Member? = jpaRepository.findByUuid(uuid)
+    override fun findByUuid(uuid: String): MemberEntity? = jpaRepository.findByUuid(uuid)
 
-    override fun findById(id: Long?): Member? {
+    override fun findById(id: Long?): MemberEntity? {
         if (id == null) return null
         return jpaRepository.findById(id).orElseGet { null }
     }
 
-    override fun save(member: Member): Member = jpaRepository.save(member)
+    override fun save(member: MemberEntity): MemberEntity = jpaRepository.save(member)
 
     override fun deleteByUuid(uuid: String) {
         jpaRepository.deleteByUuid(uuid)
@@ -36,13 +35,15 @@ open class MemberRepositoryImpl(
 
     override fun existsByUuid(uuid: String): Boolean = jpaRepository.existsByUuid(uuid)
 
-    override fun findOrCreateGuest(uuid: String, initialPoint: Long): Member = findByUuid(uuid) ?: run {
+    override fun findOrCreateGuest(uuid: String, initialPoint: Long): MemberEntity = findByUuid(uuid) ?: run {
         // Use reflection to access private constructor (uuid, initialPoint)
-        val constructor = Member::class.java.getDeclaredConstructor(String::class.java, Long::class.java)
+        val constructor = MemberEntity::class.java.getDeclaredConstructor(String::class.java, Long::class.java)
         constructor.isAccessible = true
         val guest = constructor.newInstance(uuid, initialPoint)
         save(guest)
     }
 
     override fun increasePointByUuid(uuid: String, amount: Long): Int = jpaRepository.increasePointByUuid(uuid, amount)
+
+    override fun decreasePointByUuid(uuid: String, amount: Long): Int = jpaRepository.decreasePoint(uuid, amount)
 }
